change structure&add sign up validator
