+++ README.md
@@ -17,12 +17,23 @@ $ pip install Flask
 $ pip install newsapi-python
 $ pip install pandas
 $ pip install flask-bootstrap
+$ pip install flask-sqlalchemy
+$ pip install flask-wtf
+$ pip install email-validator
+$ pip install flask-bcrypt
 ```
 Start running the application:
 ```
 $ export FLASK_APP=app.py
 $ flask run
 ```
+
+## if meet urllib error
+Run
+```
+/Applications/Python\ 3.7/Install\ Certificates.command
+```
+
 ## Authors
 
 * **Meng Ding**
