+++ app/__init__.py
@@ -1,0 +1,21 @@
+from flask import Flask
+from flask_bootstrap import Bootstrap
+from flask_bcrypt import Bcrypt
+from flask_sqlalchemy import SQLAlchemy
+
+from config import Config
+
+
+app = Flask(__name__)
+bootstrap = Bootstrap(app)
+db = SQLAlchemy(app)
+bcrypt = Bcrypt(app)
+app.config.from_object(Config)
+
+from app.routes import *
+
+
+
+
+
+
