+++ app/forms.py
@@ -1,0 +1,24 @@
+from flask_wtf import FlaskForm, RecaptchaField
+from wtforms import StringField, PasswordField, SubmitField
+from wtforms.validators import DataRequired, Length, Email, EqualTo, ValidationError
+
+from app.models import User
+
+class RegisterForm(FlaskForm):
+
+    username =StringField('Username', validators=[DataRequired(), Length(min=6, max=20)])
+    email = StringField('Email', validators=[DataRequired(), Email()])
+    password = PasswordField('password', validators=[DataRequired(), Length(min=8, max=20)])
+    confirm = PasswordField('Repeat Password', validators=[DataRequired(), EqualTo('password')])
+    recaptcha = RecaptchaField()
+    submit = SubmitField('Register')
+
+    def validate_username(self, username):
+        user = User.query.filter_by(username=username.data).first()
+        if user:
+            raise ValidationError('Username already taken. Please choose another one.')
+
+    def validate_email(self, email):
+        user = User.query.filter_by(email=email.data).first()
+        if user:
+            raise ValidationError('Email already taken.Please choose another one.')
