+++ app/templates/base.html
@@ -1,0 +1,36 @@
+{% extends "bootstrap/base.html" %}
+
+<title>
+    {% block title %}
+    {% if title %}
+     {{ title }}
+    {% else %}
+      Covid Webstie by Team 3
+    {% endif %}
+    {% endblock %}
+</title>
+
+{% block navbar %}
+  {% include 'navbar.html' %}
+{% endblock %}
+
+{% block content %}
+    <div class="container">
+        <div class="row">
+            <div class="col-md-6">
+                {% with messages = get_flashed_messages(with_categories=True) %}
+                  {% if messages %}
+                    {% for category, message in messages %}
+                      <div class="alert alert-{{ category }}">
+                          {{ message }}
+                      </div>
+                    {% endfor %}
+                  {% endif %}
+                {% endwith %}
+            </div>
+        </div>
+        {% block app_content %}
+        {% endblock %}
+    </div>
+{% endblock %}
+
