+++ app/templates/navbar.html
