+++ app/templates/register.html
@@ -1,0 +1,14 @@
+{% extends 'base.html' %}
+
+{% block app_content %}
+  <h1>Sign Up</h1>
+  <br>
+  <div class="row">
+    <div class="col-md-6">
+      {% import 'bootstrap/wtf.html' as wtf %}
+      {{ wtf.quick_form(form) }}
+    </div>
+
+  </div>
+
+{% endblock %}
