+++ requirements.txt
@@ -2,3 +2,7 @@ flask
 flask-bootstrap
 newsapi
 pandas
+flask-sqlalchemy
+flask-wtf
+email-validator
+flask-bcrypt
