--- templates/base.html
@@ -1,23 +1,0 @@
-{% extends "bootstrap/base.html" %}
-
-<title>
-    {% block title %}
-    {% if title %}
-     {{ title }}
-    {% else %}
-      Covid Webstie by Team 3
-    {% endif %}
-    {% endblock %}
-</title>
-
-{% block navbar %}
-  {% include 'navbar.html' %}
-{% endblock %}
-
-{% block content %}
-    <div class="container">
-        {% block app_content %}
-        {% endblock %}
-    </div>
-{% endblock %}
-
