--- templates/register.html
@@ -1,5 +1,0 @@
-{% extends 'base.html' %}
-
-{% block app_content %}
-  <h1>Register Now</h1>
-{% endblock %}
